feat(integrations/netzodb): create `netzodb` integration
