+++ lib/integrations/databases/netzodb.test.ts
@@ -1,0 +1,11 @@
+import { assertExists } from "../../deps/std/assert/mod.ts";
+import { netzodb } from "./netzodb.ts";
+
+const db = netzodb({ kv: await Deno.openKv(":memory:") });
+
+Deno.test("[dbs] netzodb", { ignore: false }, async (t) => {
+  await t.step("declarations", () => {
+    assertExists(netzodb);
+    assertExists(db);
+  });
+});
