+++ lib/integrations/databases/redis.ts
@@ -5,12 +5,12 @@ export type RedisOptions = {
 };
 
 /**
- * Factory function for the Redis DB
+ * Factory function for the Redis database
  *
  * @see https://netzo.io/docs/modules/databases/redis
  *
  * @param {string} url - connection string URL to the Redis instance in the form of "redis[s]://[[username][:password]@][host][:port][/db-number]"
- * @returns {object} - an DB client instance
+ * @returns {object} - a DB client instance
  */
 export const redis = async ({
   url = Deno.env.get("REDIS_URL")!,
