+++ templates/crm/netzo.config.ts
@@ -1,18 +1,15 @@
-import { createDatabase } from "netzo/database/mod.ts";
+import { netzodb } from "netzo/integrations/databases/netzodb.ts";
 import { defineConfig } from "netzo/mod.ts";
 import * as netzo from "netzo/plugins/mod.ts";
 import { unocss } from "netzo/plugins/unocss/plugin.ts";
 import unocssConfig from "./unocss.config.ts";
 
-export const kv = await Deno.openKv();
-
-export const db = createDatabase(kv);
+export const db = netzodb();
 
 export default defineConfig({
   plugins: [
     netzo.environments(),
     netzo.api({
-      kv,
       apiKey: Deno.env.get("NETZO_API_KEY"),
       collections: [
         { name: "accounts", idField: "id" },
